requirements.txt
Django>=4.2
djangorestframework
dnspython
python-whois
drf-yasg
